+++ src/aidas/vis/__init__.py
@@ -5,11 +5,14 @@ aidas.vis (aidas.vis)
 
 .. currentmodule:: aidas.vis
 
-This module handles the visualization of the lake breeze mask on radar data.
+This module handles the visualization of detection masks on radar data.
 
 .. autosummary::
     :toctree: generated/
 
     visualize_lake_breeze
+    visualize_velocity_waves
 """
 from .visualize_lake_breeze import visualize_lake_breeze
+from .visualize_velocity_waves import visualize_velocity_waves
+